refactor(database): call them creation groups, not group tags
Dario's point: "group" on its own is ambiguous and will get more so -- a chat
brings groups of its own, and Gradido already has its circles. These are the
groups a creation belongs to, hence "creation group". The word is Bernd's.

Migration 0110 renames exactly what 0108 created: 3 tables, 4 columns, 7
indices. The two columns 0109 added carry no group wording and stay. Columns
first while the tables still answer to their old names, then the tables, then
the indices through the new names; the downgrade walks the exact inverse.

★ A rename migration rather than editing 0108, although no release contains
0108 yet and production has never seen these tables: staging and every
developer database ran it long ago, and an edited migration does not run again
for them. Editing it would leave those databases on the old names under code
expecting the new ones, silently.

The indices are renamed rather than left alone on purpose -- idx_cgt_group_tag_id
sitting on a table called contribution_creation_groups is a trap for whoever
reads the schema next.
