+++ admin/src/utils/creationGroupLabel.spec.js
