+++ backend/src/graphql/resolver/CreationGroupAssignment.test.ts
