+++ backend/src/graphql/resolver/CreationGroupResolver.ts
