+++ backend/src/graphql/resolver/util/attachContributionCreationGroups.ts
