+++ backend/src/graphql/resolver/util/contributionCreationGroups.ts
