+++ backend/src/graphql/resolver/util/creationGroupsByContribution.ts
