+++ backend/src/graphql/resolver/util/moderatorCreationGroupScope.ts
