+++ backend/src/graphql/resolver/util/userCreationGroups.ts
