+++ database/src/entity/UserCreationGroup.ts
@@ -8,16 +8,16 @@ import {
   UpdateDateColumn,
 } from 'typeorm'
 
-// Group functions: a user's personal group-tag list.
+// Group functions: a user's personal creation-group list.
 // The entry with the lowest sortOrder is the user's main tag, pre-filled in the
-// group-tag field on submission. User-maintained AND moderator-editable — healing a
+// creation-group field on submission. User-maintained AND moderator-editable — healing a
 // forgotten/misspelled tag happens on the user's list, not just on one contribution.
 // No FK constraints (matches the crea_records convention) — indices only.
-@Index('uniq_user_group_tag', ['userId', 'groupTagId'], { unique: true })
-@Entity('user_group_tags', {
+@Index('uniq_user_creation_group', ['userId', 'creationGroupId'], { unique: true })
+@Entity('user_creation_groups', {
   engine: 'InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci',
 })
-export class UserGroupTag extends BaseEntity {
+export class UserCreationGroup extends BaseEntity {
   @PrimaryGeneratedColumn('increment', { unsigned: true })
   id: number
 
@@ -26,8 +26,8 @@ export class UserGroupTag extends BaseEntity {
   userId: number
 
   @Index()
-  @Column({ name: 'group_tag_id', type: 'int', unsigned: true, nullable: false })
-  groupTagId: number
+  @Column({ name: 'creation_group_id', type: 'int', unsigned: true, nullable: false })
+  creationGroupId: number
 
   @Column({ name: 'sort_order', type: 'int', unsigned: true, nullable: false, default: 0 })
   sortOrder: number
