+++ database/src/entity/index.ts
@@ -1,15 +1,15 @@
 import { Community } from './Community'
 import { CommunityHandshakeState } from './CommunityHandshakeState'
 import { Contribution } from './Contribution'
-import { ContributionGroupTag } from './ContributionGroupTag'
+import { ContributionCreationGroup } from './ContributionCreationGroup'
 import { ContributionLink } from './ContributionLink'
 import { ContributionMessage } from './ContributionMessage'
 import { CreaRecord } from './CreaRecord'
 import { CreaSetting } from './CreaSetting'
+import { CreationGroup } from './CreationGroup'
 import { DltTransaction } from './DltTransaction'
 import { Event } from './Event'
 import { FederatedCommunity } from './FederatedCommunity'
-import { GroupTag } from './GroupTag'
 import { LoginElopageBuys } from './LoginElopageBuys'
 import { Migration } from './Migration'
 import { PendingTransaction } from './PendingTransaction'
@@ -17,14 +17,14 @@ import { Transaction } from './Transaction'
 import { TransactionLink } from './TransactionLink'
 import { User } from './User'
 import { UserContact } from './UserContact'
-import { UserGroupTag } from './UserGroupTag'
+import { UserCreationGroup } from './UserCreationGroup'
 import { UserRole } from './UserRole'
 
 export {
   Community,
   CommunityHandshakeState,
   Contribution,
-  ContributionGroupTag,
+  ContributionCreationGroup,
   ContributionLink,
   ContributionMessage,
   CreaRecord,
@@ -32,7 +32,7 @@ export {
   DltTransaction,
   Event,
   FederatedCommunity,
-  GroupTag,
+  CreationGroup,
   LoginElopageBuys,
   Migration,
   PendingTransaction,
@@ -40,7 +40,7 @@ export {
   TransactionLink,
   User,
   UserContact,
-  UserGroupTag,
+  UserCreationGroup,
   UserRole,
 }
 
@@ -48,7 +48,7 @@ export const entities = [
   Community,
   CommunityHandshakeState,
   Contribution,
-  ContributionGroupTag,
+  ContributionCreationGroup,
   ContributionLink,
   ContributionMessage,
   CreaRecord,
@@ -56,7 +56,7 @@ export const entities = [
   DltTransaction,
   Event,
   FederatedCommunity,
-  GroupTag,
+  CreationGroup,
   LoginElopageBuys,
   Migration,
   PendingTransaction,
@@ -64,6 +64,6 @@ export const entities = [
   TransactionLink,
   User,
   UserContact,
-  UserGroupTag,
+  UserCreationGroup,
   UserRole,
 ]
