+++ frontend/src/utils/creationGroupLabel.js
